[Parsing] split expression module
